CLN/입력받는 기능InputValue 클래스로 분리
